Fix errant removal of some side-effecting assigns
Prior to this snapshot, some repeated assignments and assignments to a
never-read variable would get optimized away, even if the RHS of the
assignment had a side-effect like writing to one of its arguments.

Now, those statements are left intact. This change is perhaps too
conservative in that it keeps the variables live (and not just the
assignment statements). An alternative would be instead to propogate
some kind of negative value in that slot (I believe this is what LLVM
calls a "poison" value) so that other optimizations can "see" the var is
holds no real value.