Expand shortened module names when searching for files
Verilator hashes very long module names (>=128 characters) so that when it writes out the .cpp files they don't hit OS path/file length limits.

However that is transformation is done before the final output step, so we need to account for it when using module names. In this case, when auto-searching for a module named `foo` in a file called `foo.sv`. If `foo` is very long then `modname` in `filePath()` will be the shortened (hashed) version, but we actually want to search for the unshortened file name.

Fixes #5196