Use helpman's -n option to create man pages with useful whatis section
The whatis entry (the brief description found in the NAME section) for
manual pages created by help2man is of the form:

 program - manual page for program

This conveys no information about what the program is for and is
repetitive. The short description should contain brief information about
what the program is for to aid in searching with apropos and similar
programs.