Current comparison of `prd_pos` with `dot_pos` doesn't take the case of `prd_pos` being `string::npos` into account.
For example: in `TOP__DOT__top__DOT__sub`, `prd_pos == string::npos`. `string::npos` is defined to be -1, but it's unsigned (https://en.cppreference.com/w/cpp/string/basic_string/npos), meaning -1 is converted to the largest integer. This results with the `else` case, which is incorrect.

In addition, we removed the misleading comment.

Fixes #5314.
Ping @AndrewNolte.