Apply DFG before V3Gate.
This patch enables DFG to run after V3Scope (but before V3ActiveTop),
and applies it just before V3Gate.

This is beneficial because a lot of lowering and optimization has been
performed at that point (V3Task, V3Unroll, V3SplitAs), which enables
more code to convert to DFG form.

While this yields +/- a few % on most RTLMeter designs, one case gains
17% simulation speed.

More importantly I'm working on some patches to split up combinational
cycles, and the most problematic ones happen through module boundaries,
so this enables handling those.

There is a slight risk of reducing the efficiency of V3Combine, as where
some logic must be inserted when converting back to Ast is not entirely
clearing scoped mode. In this patch, we put the logic into the scope that
contains the AstVarScope that it drives.

The binToOneHot pass cannot yet handle a scoped DfgGraph, so is not
applied in the post scope run.