Optimize Shift pattern in V3Const
This pattern creates 3 operations from 2. This might have been
beneficial for expanded wide expressions if the shift was over
VL_EDATA_SIZE, but V3Subst will delete word assignments on the input in
that case. The real problem is this pattern will turn an And/Or/Xor tree
with N leaves (so between log(N) to N nodes + 1 Shift) into the same
log(N) to N nodes + N shifts, and on some RTLMeter designs this blows up
compile time and memory.

Restrict the pattern to apply only if we know that the resulting
expressions will be further reducible hence not increasing total number
of operations.