Tests: Fix driver.py --xrun option raising error
The member variable self.xrun is set to 0, causing the compile function
to not chose the "elif param['xrun']" path and instead failing with the
error message "No compile step defined for 'xrun' scenario". This
appears to be a remnant of the previous driver.pl script. This commit
removes this variable assignment and allows driver.py to launch xrun
again.

Signed-off-by: Christian Hecken <christian.hecken@ibm.com>