Internals: Make AstCExpr always cleanOut (#6280)
Roundabout patch, but the cleanOut flag in AstCExpr is a bit in the way
for cleanups necessary for #6280.

There was exactly one place in V3Task, handling DPI arguments when we
relied on cleanOut of AstCExpr being false for masking. Made that code
do the relevant masking via a few new run-time functions, which also
eliminates some special cases in the relevant V3Task functions.