Do not ignore virtual interface member triggers
Virtual interface member triggers used to not be added to the trigger
vector computation, and instead they resolved to some random trigger
bit. And the actual variables inserted in V3SchedVirIface for virtual
interface member triggers were subsequently unused and removed.

Rewrote to what the likely intention was, but this brings back the did
not converge issue reported in #5116 in two tests. Those test don't
actually test the virtual interface member trigger, they happen to pass
because the dependent combinational logic is resolved in ICO in one
iteration which is always executed at least once on the initial
iteration of ICO. (Also all combinational logic dependent on a virtual
interface is always replicated into ICO, which I'm not sure is
warranted, but it happens to make these tests "pass").