Tests: Add --gdbsim to iverilog with use_libvpi
To debug vpi with a simulator other than Verilator, the `libvpi.so`
needs to be compiled with debug symbols. This commit adds `-ggdb` to the
VPI compilation flags when `--gdbsim` is selected.
To run under gdb, the run `cmd` for 'iv' in the execute function is
modified to be prefixed with gdb if `use_libvpi = True` and `--gdbsim`
is selected.