Tests: Fix Icarus vvp execution with use_libvpi
Previously, `obj_iv/<testname>/simiv` was executed directly and run
flags were appended to it. However, when `use_libvpi = True`, the `vvp`
command was appended to it together with `libvpi.so`. This is not the
correct syntax for calling the `vvp` runtime, according to the [Icarus
documentation about using VPI]
(https://steveicarus.github.io/iverilog/usage/vpi.html), and it did
not execute correctly.
This commit fixes this behavior by prepending the `vvp` runtime, then
inserting the run flags and, if `use_libvpi = True`, also the path to
the `libvpi.so`, and finally the `simiv` at the end.